demo(source): refactor to separate modules
